+++ 14th-team5-iOS/App/Sources/Presentation/Camera/Reactor/CameraDisplayViewReactor.swift
@@ -9,6 +9,7 @@ import Foundation
 
 import Data
 import Domain
+import DesignSystem
 import ReactorKit
 import Core
 
@@ -140,7 +141,14 @@ public final class CameraDisplayViewReactor: Reactor {
                     }
             )
         case .didTapArchiveButton:
-            cameraDisplayNavigator.showArchiveToast()
+            let config = BBToastConfiguration(direction: .bottom(yOffset: -360), animationTime: 1.0)
+            let viewConfig = BBToastViewConfiguration(minWidth: 207)
+            provider.bbToastService.show(
+                image:DesignSystemAsset.camera.image.withTintColor(DesignSystemAsset.gray300.color),
+                title: "사진이 저장되었습니다.",
+                viewConfig: viewConfig,
+                config: config
+                )
             return .concat(
                 .just(.setLoading(false)),
                 .just(.saveDeviceimage(currentState.displayData)),
@@ -186,11 +194,25 @@ public final class CameraDisplayViewReactor: Reactor {
                 .just(.setDisplayEditSection([]))
             )
         case .showInputTextError:
-            cameraDisplayNavigator.showToast()
+            let config = BBToastConfiguration(direction: .bottom(yOffset: -360), animationTime: 1.0)
+            let viewConfig = BBToastViewConfiguration(minWidth: 207)
+            provider.bbToastService.show(
+                image: DesignSystemAsset.warning.image,
+                title: "8자까지 입력 가능해요",
+                viewConfig: viewConfig,
+                config: config
+            )
             return .empty()
             
         case let .showInputBlankTextError(displayText):
-            cameraDisplayNavigator.showWarningToast()
+            let config = BBToastConfiguration(direction: .bottom(yOffset: -360), animationTime: 1.0)
+            let viewConfig = BBToastViewConfiguration(minWidth: 207)
+            provider.bbToastService.show(
+                image: DesignSystemAsset.warning.image,
+                title: "띄어쓰기는 할 수 없어요",
+                viewConfig: viewConfig,
+                config: config
+            )
             let generateText = displayText.trimmingCharacters(in: .whitespaces)
             return .just(.setTrimedText(generateText))
         }
